Speedup training

+++ diffengine/configs/dreambooth/stable_diffusion_v15_dreambooth_lora_dog_pre_compute.py
@@ -6,4 +6,5 @@ with read_base():
     from .._base_.models.stable_diffusion_v15_lora import *
     from .._base_.schedules.stable_diffusion_1k import *
 
-model.update(pre_compute_text_embeddings=True)
+model.update(pre_compute_text_embeddings=True,
+             weight_dtype="bf16")
